Code Refactoring
custom hook과 function을 사용해 코드 리펙토링을 진행했습니다.
