+++ src/components/create/InputForm.jsx
@@ -1,15 +1,43 @@
 import styled from 'styled-components';
-import React from 'react';
 import { Paper } from '@mui/material';
+import { useParams, useNavigate } from 'react-router-dom';
+import React from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 import Header from './Header';
 import InputTitle from './InputTitle';
 import InputContents from './InputContents';
+import { createPost, updatePost } from '../../redux/modules/postSlice';
+import { clearText } from '../../redux/modules/textSlice';
+import newPost from '../../functions/newPost';
+import useSetUpdateText from '../../hooks/useSetUpdateText';
 
 function InputForm({ isCreate }) {
+  const { postId } = useParams();
+  const postText = useSelector((state) => state.postText); // post의 현재 title, contents, name, password의 state입니다.
+
+  const dispatch = useDispatch();
+  const navigate = useNavigate();
+
+  const onAddHandler = (event) => {
+    event.preventDefault();
+    dispatch(createPost(newPost(postText)));
+    dispatch(clearText());
+    navigate('/');
+  };
+
+  const onUpdateHandler = (event) => {
+    event.preventDefault();
+    dispatch(updatePost({ ...postText, id: postId }));
+    dispatch(clearText());
+    navigate(`/details/${postId}`);
+  };
+
+  useSetUpdateText();
+
   return (
     <StInputForm>
       <Paper sx={{ width: '100%', bgcolor: 'background.paper', p: 2, boxSizing: 'border-box' }}>
-        <Header isCreate={isCreate} />
+        <Header isCreate={isCreate} onAddHandler={onAddHandler} onUpdateHandler={onUpdateHandler} />
         <InputTitle isCreate={isCreate} />
         <InputContents isCreate={isCreate} />
       </Paper>
