+++ src/redux/modules/postSlice.jsx
@@ -1,6 +1,5 @@
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 import axios from 'axios';
-import Moment from 'react-moment';
 
 const initialState = {
   posts: [],
